+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * [#151](https://github.com/exercism/cli/pull/151): Expand '~' in config path to home directory - @lcowell
 * Stop supporting legacy config files (~/.exercism.go)
+* Deleted deprecated login/logout commands
 * Your contribution here
 
 ## v1.9.2 (Jan 11, 2015)